Work around deprecation of update_attribute in AR 3.2.7
The update_attribute method has been deprecated in ActiveRecord 3.2.7
and will be removed in ActiveRecord 4.0.0. Work around this in two
ways, first by creating instance methods where it's used in models
and second by using update_column in specs. Where update_column is not
available, update_attribute is used as a fallback.