Remove duplicate ActiveRecord dependency
The gemspec defines ActiveRecord as a dependency twice, the second of which
to ensure that a compatible version is available for development. Migrating
this version lock to the Gemfile results in the same behaviour, without the
inconvenience of Bundler >= v1.10.1 refusing to install the gem due to an
invalid gemspec.

An example of this error message follows:
```
/opt/rubies/2.2.2/gems/bundler-1.10.3/lib/bundler.rb:374:in `rescue in load_gemspec_uncached':
The gemspec at /www/sites/demo/vendor/bundle/ruby/2.2.2/bundler/gems/oauth2-provider-b60ea985a0e9/songkick-oauth2-provider.gemspec
is not valid. The validation error was 'duplicate dependency on activerecord (~> 3.2.0, development), (>= 0) use: (Bundler::InvalidOption)
    add_runtime_dependency 'activerecord', '~> 3.2.0', '>= 0'
```